+++ APLC_plugin/Locations.cs
@@ -204,9 +204,14 @@ public class ScrapLocations : Locations
      */
     public override void CheckComplete()
     {
+        GameObject cruiser = GameObject.FindObjectsByType<VehicleController>(sortMode: FindObjectsSortMode.None).FirstOrDefault(vehicle => vehicle.magnetedToShip)?.gameObject;
+        bool hasCruiser = cruiser != null;
+
         var list = (from obj in GameObject.Find("/Environment/HangarShip").GetComponentsInChildren<GrabbableObject>()
-            where obj.name != "ClipboardManual" && obj.name != "StickyNoteItem"
-            select obj).ToList();
+                    where obj.name != "ClipboardManual" && obj.name != "StickyNoteItem"
+                    select obj).Union(hasCruiser ? (from obj in cruiser.GetComponentsInChildren<GrabbableObject>()
+                                                    where obj.name != "CompanyCruiserManual(Clone)"
+                                                    select obj) : []).ToList();
         foreach (var scrap in list)
         {
             string scrapName = scrap.itemProperties.itemName;